Move subgraph primitives.
