+++ pdr_backend/analytics/test/test_get_traction_info.py
@@ -5,9 +5,9 @@ import pytest
 from enforce_typing import enforce_types
 
 from pdr_backend.analytics.get_predictions_info import get_traction_info_main
+from pdr_backend.lake.prediction import Prediction
 from pdr_backend.lake.table import Table
 from pdr_backend.ppss.ppss import mock_ppss
-from pdr_backend.subgraph.prediction import Prediction
 
 
 @enforce_types
