+++ pdr_backend/lake/gql_data_factory.py
@@ -5,15 +5,17 @@ import polars as pl
 from enforce_typing import enforce_types
 
 from pdr_backend.lake.csv_data_store import CSVDataStore
+from pdr_backend.lake.payout import Payout
 from pdr_backend.lake.persistent_data_store import PersistentDataStore
 from pdr_backend.lake.plutil import _object_list_to_df
+from pdr_backend.lake.prediction import Prediction
+from pdr_backend.lake.slot import Slot
+from pdr_backend.lake.subscription import Subscription
 from pdr_backend.lake.table import Table, TableType, TempTable, get_table_name
 from pdr_backend.lake.table_pdr_predictions import _transform_timestamp_to_ms
 from pdr_backend.lake.table_registry import TableRegistry
+from pdr_backend.lake.trueval import Trueval
 from pdr_backend.ppss.ppss import PPSS
-from pdr_backend.subgraph.payout import Payout
-from pdr_backend.subgraph.prediction import Prediction
-from pdr_backend.subgraph.slot import Slot
 from pdr_backend.subgraph.subgraph_payout import fetch_payouts
 from pdr_backend.subgraph.subgraph_predictions import (
     fetch_filtered_predictions,
@@ -22,8 +24,6 @@ from pdr_backend.subgraph.subgraph_predictions import (
 from pdr_backend.subgraph.subgraph_slot import fetch_slots
 from pdr_backend.subgraph.subgraph_subscriptions import fetch_filtered_subscriptions
 from pdr_backend.subgraph.subgraph_trueval import fetch_truevals
-from pdr_backend.subgraph.subscription import Subscription
-from pdr_backend.subgraph.trueval import Trueval
 from pdr_backend.util.networkutil import get_sapphire_postfix
 from pdr_backend.util.time_types import UnixTimeMs
