+++ pdr_backend/lake/test/test_gql_data_factory.py
@@ -2,10 +2,10 @@ from unittest.mock import MagicMock, patch
 
 from pdr_backend.lake.gql_data_factory import GQLDataFactory
 from pdr_backend.lake.persistent_data_store import PersistentDataStore
+from pdr_backend.lake.prediction import mock_daily_predictions
 from pdr_backend.lake.table import TableType, get_table_name
 from pdr_backend.lake.table_registry import TableRegistry
 from pdr_backend.ppss.ppss import mock_ppss
-from pdr_backend.subgraph.prediction import mock_daily_predictions
 from pdr_backend.util.time_types import UnixTimeMs
