+++ pdr_backend/subgraph/subgraph_subscriptions.py
@@ -4,7 +4,7 @@ from typing import List
 
 from enforce_typing import enforce_types
 
-from pdr_backend.subgraph.subscription import Subscription
+from pdr_backend.lake.subscription import Subscription
 from pdr_backend.subgraph.core_subgraph import query_subgraph
 from pdr_backend.subgraph.info725 import info725_to_info
 from pdr_backend.util.networkutil import get_subgraph_url
